+++ languages/el.php
@@ -1,27 +1,11 @@
 <?php
 /**
- *   __ _  ___  ___ ___   ___   ___     ____ _ __ ___   ___
- *  / _` |/  / / __/ _ \ / _ \ /  /    / __/| '_ ` _ \ /  /
- * | (_| |\  \| (_| (_) | (_) |\  \   | (__ | | | | | |\  \
- *  \__,_|/__/ \___\___/ \___/ /__/    \___\|_| |_| |_|/__/
- * 
- * 
- ************************************************************************************
- * @ASCOOS-NAME        	: ASCOOS CMS 24'                                            *
- * @ASCOOS-VERSION     	: 24.0.0                                                    *
- * @ASCOOS-CATEGORY    	: Block (Frontend and Administrator Side)                   *
- * @ASCOOS-CREATOR     	: Drogidis Christos                                         *
- * @ASCOOS-SITE        	: www.ascoos.com                                            *
- * @ASCOOS-LICENSE     	: [Commercial] http://docs.ascoos.com/lics/ascoos/AGL.html  *
- * @ASCOOS-COPYRIGHT   	: Copyright (c) 2007 - 2024, AlexSoft Software.             *
- ************************************************************************************
- *
  * @package            	: Latest classes from PHPClasses.org
  * @subpackage         	: Greek Language file
- * @source             	: /latest-phpclasses-packages/languages/el.php
- * @version            	: 1.0.0
+ * @source             	: /phpclasses-feed/languages/el.php
+ * @version            	: 1.0.1
  * @created            	: 2024-10-14 07:00:00 UTC+3
- * @updated            	: 
+ * @updated            	: 2024-10-15 07:00:00 UTC+3
  * @author             	: Drogidis Christos
  * @authorSite         	: www.alexsoft.gr
  * @license 			: AGL-F
@@ -32,12 +16,9 @@
  * @since PHP 5.6.40
  */
 
-
-defined ("ALEXSOFT_RUN_CMS") or die("Prohibition of Access.");
-
 class TPHPClasses_Language
 {
-	public $title = "Πρόσφατες κλάσεις από το PHPClasses.org";
+	public $title = "Πρόσφατα από το PHPClasses";
 	public $more = "περισσότερα";
 }
 ?>