+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## [1.0.1] - 2024-10-16
+
+- Renamed from `latest-phpclasses-packages` to `phpclasses-feed`
+- Fixed all codes for new package name
+
+---
+
 ## [1.0.0] - 2024-10-14
 
 - Initial release